Backend/requirements.txt
fastapi
uvicorn
pydantic
tweepy
google-generativeai
sqlalchemy
python-dotenv